+++ secretflow/ml/nn/fl/backend/torch/strategy/fed_pac.py
@@ -15,8 +15,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from secretflow.ml.nn.fl.backend.torch.fl_base import FedPACTorchModel
+from secretflow.ml.nn.fl.backend.torch.fl_base import BaseTorchModel
 from secretflow.ml.nn.fl.strategy_dispatcher import register_strategy
+from secretflow.ml.nn.core.torch import BuilderType
 
 from copy import deepcopy
 import torch
@@ -24,9 +25,195 @@ from typing import Dict, Tuple
 import copy
 import numpy as np
 import logging
+from torch import nn
 
+class FedPAC(BaseTorchModel):
+    def __init__(
+        self,
+        builder_base: BuilderType,
+        random_seed: int = None,
+        skip_bn: bool = False,
+        **kwargs,
+    ):
+        super().__init__(builder_base, random_seed=random_seed, **kwargs)
+        self.num_classes = kwargs.get("num_classes", 10)
+        self.criterion = nn.CrossEntropyLoss()
+        self.local_model = self.model
+        self.w_local_keys = self.local_model.classifier_weight_keys
+        self.local_ep_rep = 1
+        self.global_protos = {}
+        self.g_protos = None
+        self.mse_loss = nn.MSELoss()
+        self.lam = kwargs.get("lam", 1.0)  # 1.0 for mse_loss
+    
+    def prior_label(self, dataset):
+        py = torch.zeros(self.num_classes)
+        total = len(dataset.dataset)
+        data_loader = iter(dataset)
+        iter_num = len(data_loader)
+        for it in range(iter_num):
+            images, labels = next(data_loader)
+            for i in range(self.num_classes):
+                py[i] = py[i] + (i == labels).sum()
+        py = py / (total)
+        return py
 
-class FedPAC(FedPACTorchModel):
+    def size_label(self, dataset):
+        py = torch.zeros(self.num_classes)
+        total = len(dataset.dataset)
+        data_loader = iter(dataset)
+        iter_num = len(data_loader)
+        for it in range(iter_num):
+            images, labels = next(data_loader)
+            for i in range(self.num_classes):
+                py[i] = py[i] + (i == labels).sum()
+
+        py = py / (total)
+        size_label = py * total
+        return size_label
+
+    def sample_number(self):
+        data_size = len(self.train_set.dataset)
+        w = torch.tensor(data_size).to(self.exe_device)
+        return w
+
+    def evaluate(self, step_per_epoch=0) -> Tuple[float, float]:
+        assert self.model is not None, "Model cannot be none, please give model define"
+        assert (
+            len(self.model.metrics) > 0
+        ), "Metric cannot be none, please give metric by 'TorchModel'"
+        self.model.eval()
+        model = self.local_model
+        device = self.exe_device
+        correct = 0
+        eval_loader = self.eval_set
+        loss_test = []
+        self.reset_metrics()
+        with torch.no_grad():
+            for inputs, labels in eval_loader:
+                inputs, labels = inputs.to(device), labels.to(device)
+                _, outputs = model(inputs)
+                model.update_metrics(outputs, labels)
+                loss = self.criterion(outputs, labels)
+                loss_test.append(loss.item())
+                _, predicted = torch.max(outputs.data, 1)
+                correct += (predicted == labels).sum().item()
+            result = {}
+            self.transform_metrics(result, stage="eval")
+
+        if self.logs is None:
+            self.wrapped_metrics.extend(self.wrap_local_metrics())
+            return self.wrap_local_metrics()
+        else:
+            val_result = {}
+            for k, v in result.items():
+                val_result[f"val_{k}"] = v
+            self.logs.update(val_result)
+            self.wrapped_metrics.extend(self.wrap_local_metrics(stage="val"))
+            return self.wrap_local_metrics(stage="val")
+
+    def get_local_protos(self, step, train_steps):
+        model = self.local_model
+        local_protos_list = {}
+        step_counter = 0
+        total_steps = 0
+        for inputs, labels in self.train_set:
+            if total_steps < step:
+                total_steps += 1
+                continue
+            if step_counter >= train_steps:
+                break
+            inputs, labels = inputs.to(self.exe_device), labels.to(self.exe_device)
+            features, outputs = model(inputs)
+            protos = features.clone().detach()
+            for i in range(len(labels)):
+                if labels[i].item() in local_protos_list.keys():
+                    local_protos_list[labels[i].item()].append(protos[i, :])
+                else:
+                    local_protos_list[labels[i].item()] = [protos[i, :]]
+            step_counter += 1
+            total_steps += 1
+
+        local_protos = {}
+        for [label, proto_list] in local_protos_list.items():
+            proto = 0 * proto_list[0]
+            for p in proto_list:
+                proto += p
+            local_protos[label] = proto / len(proto_list)
+        return local_protos
+
+    def get_local_protos_with_entire_dataset(self):
+        model = self.local_model
+        local_protos_list = {}
+        for inputs, labels in self.train_set:
+            inputs, labels = inputs.to(self.exe_device), labels.to(self.exe_device)
+            features, outputs = model(inputs)
+            protos = features.clone().detach()
+            for i in range(len(labels)):
+                if labels[i].item() in local_protos_list.keys():
+                    local_protos_list[labels[i].item()].append(protos[i, :])
+                else:
+                    local_protos_list[labels[i].item()] = [protos[i, :]]
+
+        local_protos = {}
+        for [label, proto_list] in local_protos_list.items():
+            proto = 0 * proto_list[0]
+            for p in proto_list:
+                proto += p
+            local_protos[label] = proto / len(proto_list)
+        return local_protos
+
+    def statistics_extraction(self):
+        model = self.local_model
+        cls_keys = self.w_local_keys
+        g_params = (
+            model.state_dict()[cls_keys[0]]
+            if isinstance(cls_keys, list)
+            else model.state_dict()[cls_keys]
+        )
+        d = g_params[0].shape[0]
+        feature_dict = {}
+        datasize = torch.tensor(len(self.train_set.dataset)).to(self.exe_device)
+        with torch.no_grad():
+            for inputs, labels in self.train_set:
+                inputs, labels = inputs.to(self.exe_device), labels.to(self.exe_device)
+                features, outputs = model(inputs)
+                feat_batch = features.clone().detach()
+                for i in range(len(labels)):
+                    yi = labels[i].item()
+                    if yi in feature_dict.keys():
+                        feature_dict[yi].append(feat_batch[i, :])
+                    else:
+                        feature_dict[yi] = [feat_batch[i, :]]
+        for k in feature_dict.keys():
+            feature_dict[k] = torch.stack(feature_dict[k])
+
+        py = self.prior_label(self.train_set).to(self.exe_device)
+        py2 = py.mul(py)
+        v = 0
+        h_ref = torch.zeros((self.num_classes, d), device=self.exe_device)
+        for k in range(self.num_classes):
+            if k in feature_dict.keys():
+                feat_k = feature_dict[k]
+                num_k = feat_k.shape[0]
+                feat_k_mu = feat_k.mean(dim=0)
+                h_ref[k] = py[k] * feat_k_mu
+                v += (
+                    py[k] * torch.trace((torch.mm(torch.t(feat_k), feat_k) / num_k))
+                ).item()
+                v -= (py2[k] * (torch.mul(feat_k_mu, feat_k_mu))).sum().item()
+        v = v / datasize.item()
+        return v, h_ref
+
+    def get_statistics(
+        self,
+    ) -> Tuple[float, torch.Tensor, torch.Tensor, torch.Tensor, list, torch.Tensor]:
+        v, h_ref = self.statistics_extraction()
+        label_size = self.size_label(self.train_set).to(self.exe_device)
+        sample_num = self.sample_number()
+        dataset_size = torch.tensor(len(self.train_set.dataset)).to(self.exe_device)
+        return v, h_ref, label_size, dataset_size, self.w_local_keys, sample_num
+
     def train_step(
         self,
         step,
@@ -40,7 +227,8 @@ class FedPAC(FedPACTorchModel):
         """Accept ps model params, then do local train
 
         Args:
-            cur_steps: current train step
+            step: current train step
+            cur_steps: current train step in the whole training process
             train_steps: local training steps
             kwargs: strategy-specific parameters
         Returns:
@@ -57,7 +245,6 @@ class FedPAC(FedPACTorchModel):
         iter_loss = []
         model.zero_grad()
         global_protos = self.global_protos
-        self.last_model = deepcopy(model)
         # get local prototypes before training, dict:={label: list of sample features}
         local_protos1 = self.get_local_protos(step, train_steps)
         # Set optimizer for the local updates, default sgd
@@ -114,10 +301,7 @@ class FedPAC(FedPACTorchModel):
                     protos_new[i] = global_protos[yi].detach()
                 elif yi in local_protos1:
                     protos_new[i] = local_protos1[yi].detach()
-                else:
-                    logging.info(
-                        f"Key {yi} not found in both global_protos and local_protos1"
-                    )
+
             loss1 = self.mse_loss(protos_new, protos)
             loss = loss0 + self.lam * loss1
             loss.backward()
@@ -125,7 +309,6 @@ class FedPAC(FedPACTorchModel):
             iter_loss.append(loss.item())
 
         train_loss = sum(iter_loss) / len(iter_loss)
-        logging.info(f'local train loss: {train_loss}')
         logs['train-loss'] = train_loss
         self.logs = self.transform_metrics(logs)
         self.wrapped_metrics.extend(self.wrap_local_metrics())
@@ -173,9 +356,6 @@ class FedPAC(FedPACTorchModel):
                 if i in global_protos:
                     g_classes.append(torch.tensor(i))
                     g_protos.append(global_protos[i])
-                else:
-                    logging.info(f"Key {i} not found in global_protos")
-                    logging.info(f'global protos type: {type(global_protos)}')
             self.g_classes = torch.stack(g_classes).to(self.exe_device)
             self.g_protos = torch.stack(g_protos)
